made a theme-component
