+++ components/Johannes.js
@@ -1,63 +1,15 @@
-import React, { useRef, useState } from "react";
-import { View, ScrollView, ActivityIndicator, Animated } from "react-native";
+import React, { useState } from "react";
+import { View, ScrollView, ActivityIndicator } from "react-native";
 import { JOHANNES_URL } from "../utils/urls";
-import styled from "styled-components/native";
-
-const ArtistBox = styled.View`
-  flex: 1;
-  align-items: center;
-  justify-content: center;
-  background-color: rgba(0, 0, 0, 0.8);
-`;
-
-const FadeBox = (props) => {
-  const fadeAnim = useRef(new Animated.Value(0)).current;
-
-  React.useEffect(() => {
-    Animated.timing(fadeAnim, {
-      toValue: 1,
-      useNativeDriver: true,
-      delay: 500,
-    }).start();
-  }, [fadeAnim]);
-
-  return (
-    <Animated.View style={{ ...props.style, opacity: fadeAnim }}>
-      {props.children}
-    </Animated.View>
-  );
-};
-
-const BioText = styled.Text`
-  color: rgba(232, 209, 78, 1);
-  padding: 20px;
-  font-size: 17px;
-`;
-
-const ArtistImg = styled.Image`
-  width: 250px;
-  height: 250px;
-`;
-
-const TitleText = styled.Text`
-  color: rgba(232, 209, 78, 1);
-  font-style: italic;
-  font-weight: bold;
-  text-align: center;
-  width: 350px;
-  padding: 10px;
-`;
-
-const ArtisButton = styled.TouchableOpacity`
-  background: rgba(232, 209, 78, 1);
-  padding: 7px;
-`;
-
-const ButtonText = styled.Text`
-  text-align: center;
-  color: rgba(0, 0, 0, 0.8);
-  font-weight: bold;
-`;
+import {
+  ArtistBox,
+  FadeBox,
+  BioText,
+  ArtistImg,
+  TitleText,
+  ArtistButton,
+  ButtonText,
+} from "./Theme";
 
 const Johannes = () => {
   const [data, setData] = useState({});
@@ -99,9 +51,9 @@ const Johannes = () => {
           </BioText>
         </FadeBox>
         {showButton && (
-          <ArtisButton onPress={getJohannes}>
+          <ArtistButton onPress={getJohannes}>
             <ButtonText>SEE HIS WORK</ButtonText>
-          </ArtisButton>
+          </ArtistButton>
         )}
         {data.length > 0 &&
           data
