+++ components/Theme.js
@@ -1,0 +1,59 @@
+import React, { useRef } from "react";
+import { Animated } from "react-native";
+import styled from "styled-components/native";
+
+export const ArtistBox = styled.View`
+  flex: 1;
+  align-items: center;
+  justify-content: center;
+  background-color: rgba(0, 0, 0, 0.8);
+`;
+
+export const FadeBox = (props) => {
+  const fadeAnim = useRef(new Animated.Value(0)).current;
+
+  React.useEffect(() => {
+    Animated.timing(fadeAnim, {
+      toValue: 1,
+      useNativeDriver: true,
+      delay: 500,
+    }).start();
+  }, [fadeAnim]);
+
+  return (
+    <Animated.View style={{ ...props.style, opacity: fadeAnim }}>
+      {props.children}
+    </Animated.View>
+  );
+};
+
+export const BioText = styled.Text`
+  color: rgba(232, 209, 78, 1);
+  padding: 20px;
+  font-size: 17px;
+`;
+
+export const ArtistImg = styled.Image`
+  width: 250px;
+  height: 250px;
+`;
+
+export const TitleText = styled.Text`
+  color: rgba(232, 209, 78, 1);
+  font-style: italic;
+  font-weight: bold;
+  text-align: center;
+  width: 350px;
+  padding: 10px;
+`;
+
+export const ArtistButton = styled.TouchableOpacity`
+  background: rgba(232, 209, 78, 1);
+  padding: 7px;
+`;
+
+export const ButtonText = styled.Text`
+  text-align: center;
+  color: rgba(0, 0, 0, 0.8);
+  font-weight: bold;
+`;
